Mobile UI draft
I have created 4 styles for mobile UI(A,B,C,D).

Style A is said to be booklet style, but is aborted during designing
process.

Style B is what I called modern style, which its basic idea is to have
cards with correspond functions. Additionally its user interaction is by
swiping and draging the cards.

Style C is the icon style which is much similar to phone interface with
all those buttons and its correspond functionality icons. The User
interactions will be touching the icons to access different faces of the
application.

Style D is not a complete draft but is confirmed from all team members
from the mobile team to be the UI for this application. wWth a basic
idea of simple and also combining ideas from syle C.

The main reason of choosing style D is because in the previous design I
seperated functionalities into their  respective segments. So the user
heads to different pages for diffeerent useage. However it seems a
little bit too much for mobile device so I tried to join everthing in
the same page, for instance the timetable in style D, instead of go to
timetable page to check the course, its icons are built on the side of
the main page and shows the time of the course when user press on that
date.